fix bugs

+++ Modules/Oje.Section.WebMain/Interfaces/IPageLeftRightDesignItemService.cs
@@ -1,10 +1,5 @@
 ﻿using Oje.Infrastructure.Models;
 using Oje.Section.WebMain.Models.View;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Oje.Section.WebMain.Interfaces
 {
@@ -12,7 +7,7 @@ namespace Oje.Section.WebMain.Interfaces
     {
         ApiResult Create(PageLeftRightDesignItemCreateUpdateVM input, int? siteSettingId);
         ApiResult Update(PageLeftRightDesignItemCreateUpdateVM input, int? siteSettingId);
-        PageLeftRightDesignItemCreateUpdateVM GetById(int? id, int? siteSettingId);
+        PageLeftRightDesignItemCreateUpdateVM GetById(long? id, int? siteSettingId);
         ApiResult Delete(long? id, int? siteSettingId);
         GridResultVM<PageLeftRightDesignItemMainGridResultVM> GetList(PageLeftRightDesignItemMainGrid searchInput, int? siteSettingId);
     }
